further refactor, sketch out data cleaning, prep to do modeling
